Program to check whether the linked list is circular or not
Given a circular linked list, implement an algorithm which returns the
node at
the beginning of the loop.
DEFINITION
Circular linked list: A (corrupt) linked list in which a node's next
pointer points
to an earlier node, so as to make a loop in the linked list.
EXAMPLE
Input: A - > B - > C - > D - > E - > C [the same C as earlier]
Output: C